Added more commenting, added a new test for history
