+++ BrowseSharp/BrowseSharp.csproj
@@ -35,7 +35,4 @@
     <PackageReference Include="NETStandard.Library" Version="2.0.3" />
     <PackageReference Include="RestSharp" Version="106.3.1" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Document\" />
-  </ItemGroup>
 </Project>
