+++ BrowseSharp/Browser.cs
@@ -10,7 +10,7 @@ using RestSharp;
 namespace BrowseSharp
 {
     /// <summary>
-    /// Headless browser implimentation that creates documents for each web request.
+    /// Headless browser implementation that creates documents for each web request.
     /// </summary>
     public class Browser : BrowserCore, IBrowser, IBrowserTyped, IBrowserHistorySync
     {
@@ -19,20 +19,20 @@ namespace BrowseSharp
         /// </summary>
         public Browser(): base()
         {
-            _browserStandard = new StandardCore(JavascriptEngine, StyleEngine, _restClient, _restClient.CookieContainer, _history, StyleScrapingEnabled, JavascriptScrapingEnabled, DefaultUriProtocol);
-            _browserTyped = new TypedCore(JavascriptEngine, StyleEngine, _restClient, _restClient.CookieContainer, _history, StyleScrapingEnabled, JavascriptScrapingEnabled, DefaultUriProtocol);
+            _browserStandard = new StandardCore(base.JavascriptEngine, base.StyleEngine, _restClient, _restClient.CookieContainer, _history, base.StyleScrapingEnabled, base.JavascriptScrapingEnabled, DefaultUriProtocol);
+            _browserTyped = new TypedCore(base.JavascriptEngine, base.StyleEngine, _restClient, _restClient.CookieContainer, _history, base.StyleScrapingEnabled, base.JavascriptScrapingEnabled, DefaultUriProtocol);
 
         }
 
         /// <summary>
         /// Standard browser client for making standard web requests
         /// </summary>
-        StandardCore _browserStandard;
+        private StandardCore _browserStandard;
 
         /// <summary>
         /// Browser for making typed requests, deserializes response body to typed objects
         /// </summary>
-        TypedCore _browserTyped;
+        private TypedCore _browserTyped;
 
         /// <summary>
         /// Enables or disables javascript scraping on each request, sets for all browsers
@@ -58,364 +58,201 @@ namespace BrowseSharp
             }
         }
 
-        /// <summary>
-        /// Executes a request
-        /// </summary>
-        /// <param name="request"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public IDocument Execute(IRestRequest request)
         {
             return _browserStandard.Execute(request);
         }
 
-        /// <summary>
-        /// Executes get request
-        /// </summary>
-        /// <param name="request"></param>
-        /// <param name="httpMethod"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public IDocument ExecuteAsGet(IRestRequest request, string httpMethod)
         {
             return _browserStandard.ExecuteAsGet(request, httpMethod);
         }
 
-        /// <summary>
-        /// Executes post request
-        /// </summary>
-        /// <param name="request"></param>
-        /// <param name="httpMethod"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public IDocument ExecuteAsPost(IRestRequest request, string httpMethod)
         {
             return _browserStandard.ExecuteAsPost(request, httpMethod);
         }
 
-        /// <summary>
-        /// Executes request asynchronously
-        /// </summary>
-        /// <param name="request"></param>
-        /// <param name="token"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task<IDocument> ExecuteTaskAsync(IRestRequest request, CancellationToken token)
         {
             return await _browserStandard.ExecuteTaskAsync(request, token);
         }
 
-        /// <summary>
-        /// Executes request asynchronously
-        /// </summary>
-        /// <param name="request"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task<IDocument> ExecuteTaskAsync(IRestRequest request)
         {
             return await _browserStandard.ExecuteTaskAsync(request);
         }
 
-        /// <summary>
-        /// Executes get request asynchronously
-        /// </summary>
-        /// <param name="request"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task<IDocument> ExecuteGetTaskAsync(IRestRequest request)
         {
             return await _browserStandard.ExecuteGetTaskAsync(request);
         }
 
-        /// <summary>
-        /// Executes get request asynchronously
-        /// </summary>
-        /// <param name="request"></param>
-        /// <param name="token"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task<IDocument> ExecuteGetTaskAsync(IRestRequest request, CancellationToken token)
         {
             return await _browserStandard.ExecuteGetTaskAsync(request, token);
         }
 
-        /// <summary>
-        /// Executes post request asynchronously
-        /// </summary>
-        /// <param name="request"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task<IDocument> ExecutePostTaskAsync(IRestRequest request)
         {
             return await _browserStandard.ExecutePostTaskAsync(request);
         }
 
-        /// <summary>
-        /// Executes post request asynchronously
-        /// </summary>
-        /// <param name="request"></param>
-        /// <param name="token"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task<IDocument> ExecutePostTaskAsync(IRestRequest request, CancellationToken token)
         {
             return await _browserStandard.ExecutePostTaskAsync(request, token);
         }
 
-        /// <summary>
-        /// Performs get request
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public IDocument Navigate(string uri)
         {
             return _browserStandard.Navigate(uri);
         }
 
-        /// <summary>
-        /// Performs get request
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public IDocument Navigate(Uri uri)
         {
             return _browserStandard.Navigate(uri);
         }
 
-        /// <summary>
-        /// Performs get request
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <param name="headers"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public IDocument Navigate(string uri, Dictionary<string, string> headers)
         {
             return _browserStandard.Navigate(uri, headers);
         }
 
-        /// <summary>
-        /// Performs get request
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <param name="headers"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public IDocument Navigate(Uri uri, Dictionary<string, string> headers)
         {
             return _browserStandard.Navigate(uri, headers);
         }
 
-        /// <summary>
-        /// Performs get request
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <param name="headers"></param>
-        /// <param name="formData"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public IDocument Navigate(string uri, Dictionary<string, string> headers, Dictionary<string, string> formData)
         {
             return _browserStandard.Navigate(uri, headers, formData);
         }
 
-        /// <summary>
-        /// Performs get request
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <param name="headers"></param>
-        /// <param name="formData"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public IDocument Navigate(Uri uri, Dictionary<string, string> headers, Dictionary<string, string> formData)
         {
             return _browserStandard.Navigate(uri, headers, formData);
         }
 
-        /// <summary>
-        /// Performs get request asynchronously
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task<IDocument> NavigateAsync(string uri)
         {
             return await _browserStandard.NavigateAsync(uri);
         }
 
-        /// <summary>
-        /// Performs get request asynchronously
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task<IDocument> NavigateAsync(Uri uri)
         {
             return await _browserStandard.NavigateAsync(uri);
         }
 
-        /// <summary>
-        /// Performs get request asynchronously
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <param name="headers"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task<IDocument> NavigateAsync(string uri, Dictionary<string, string> headers)
         {
             return await _browserStandard.NavigateAsync(uri, headers);
         }
 
-        /// <summary>
-        /// Performs get request asynchronously
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <param name="headers"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task<IDocument> NavigateAsync(Uri uri, Dictionary<string, string> headers)
         {
             return await _browserStandard.NavigateAsync(uri, headers);
         }
 
-        /// <summary>
-        /// Performs get request asynchronously
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <param name="headers"></param>
-        /// <param name="formData"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task<IDocument> NavigateAsync(string uri, Dictionary<string, string> headers,
             Dictionary<string, string> formData)
         {
             return await _browserStandard.NavigateAsync(uri, headers, formData);
         }
 
-        /// <summary>
-        /// Performs get request asynchronously
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <param name="headers"></param>
-        /// <param name="formData"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task<IDocument> NavigateAsync(Uri uri, Dictionary<string, string> headers,
             Dictionary<string, string> formData)
         {
             return await _browserStandard.NavigateAsync(uri, headers, formData);
         }
 
-        /// <summary>
-        /// Performs a post request
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public IDocument Submit(string uri)
         {
             return _browserStandard.Submit(uri);
         }
 
-        /// <summary>
-        /// Performs a post request
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public IDocument Submit(Uri uri)
         {
             return _browserStandard.Submit(uri);
         }
 
-        /// <summary>
-        /// Performs a post request
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <param name="formData"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public IDocument Submit(string uri, Dictionary<string, string> formData)
         {
             return _browserStandard.Submit(uri, formData);
         }
 
-        /// <summary>
-        /// Performs a post request
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <param name="formData"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public IDocument Submit(Uri uri, Dictionary<string, string> formData)
         {
             return _browserStandard.Submit(uri, formData);
         }
 
-        /// <summary>
-        /// Performs a post request
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <param name="formData"></param>
-        /// <param name="headers"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public IDocument Submit(string uri, Dictionary<string, string> formData, Dictionary<string, string> headers)
         {
             return _browserStandard.Submit(uri, formData, headers);
         }
 
-        /// <summary>
-        /// Performs a post request
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <param name="formData"></param>
-        /// <param name="headers"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public IDocument Submit(Uri uri, Dictionary<string, string> formData, Dictionary<string, string> headers)
         {
             return _browserStandard.Submit(uri, formData, headers);
         }
 
-        
-
-        /// <summary>
-        /// Performs a post request asynchronously
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task<IDocument> SubmitAsync(string uri)
         {
             return await _browserStandard.SubmitAsync(uri);
         }
 
-        /// <summary>
-        /// Performs a post request asynchronously
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task<IDocument> SubmitAsync(Uri uri)
         {
             return await _browserStandard.SubmitAsync(uri);
         }
 
-        /// <summary>
-        /// Performs a post request asynchronously
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <param name="formData"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task<IDocument> SubmitAsync(string uri, Dictionary<string, string> formData)
         {
             return await _browserStandard.SubmitAsync(uri, formData);
         }
 
-        /// <summary>
-        /// Performs a post request asynchronously
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <param name="formData"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task<IDocument> SubmitAsync(Uri uri, Dictionary<string, string> formData)
         {
             return await _browserStandard.SubmitAsync(uri, formData);
         }
 
-        /// <summary>
-        /// Performs a post request asynchronously
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <param name="formData"></param>
-        /// <param name="headers"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task<IDocument> SubmitAsync(string uri, Dictionary<string, string> formData, Dictionary<string, string> headers)
         {
             return await _browserStandard.SubmitAsync(uri, formData, headers);
         }
 
-        /// <summary>
-        /// Performs a post request asynchronously
-        /// </summary>
-        /// <param name="uri"></param>
-        /// <param name="formData"></param>
-        /// <param name="headers"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task<IDocument> SubmitAsync(Uri uri, Dictionary<string, string> formData, Dictionary<string, string> headers)
         {
             return await _browserStandard.SubmitAsync(uri, formData, headers);
@@ -586,218 +423,241 @@ namespace BrowseSharp
             return _browserStandard.SubmitForm(form);
         }
 
-        /// <summary>
-        /// Submits form
-        /// </summary>
-        /// <param name="form"></param>
-        /// <param name="headers"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public IDocument SubmitForm(Form form, Dictionary<string, string> headers)
         {
             return _browserStandard.SubmitForm(form,headers);
         }
         
-        /// <summary>
-        /// Submits a form asynchronously
-        /// </summary>
-        /// <param name="form"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task<IDocument> SubmitFormAsync(Form form)
         {
             return await _browserStandard.SubmitFormAsync(form);
         }
 
-        /// <summary>
-        /// Submits a form asynchronously
-        /// </summary>
-        /// <param name="form"></param>
-        /// <param name="headers"></param>
-        /// <returns></returns>
+        /// <inheritdoc />
         public async Task<IDocument> SubmitFormAsync(Form form, Dictionary<string, string> headers)
         {
             return await _browserStandard.SubmitFormAsync(form, headers);
         }
-
+        
+        /// <inheritdoc />
         public IDocument<T> Execute<T>(IRestRequest request)
         {
             return _browserTyped.Execute<T>(request);
         }
-
+        
+        /// <inheritdoc />
         public IDocument<T> ExecuteAsGet<T>(IRestRequest request, string httpMethod)
         {
             return _browserTyped.ExecuteAsGet<T>(request, httpMethod);
         }
 
+        /// <inheritdoc />
         public IDocument<T> ExecuteAsPost<T>(IRestRequest request, string httpMethod)
         {
             return _browserTyped.ExecuteAsPost<T>(request, httpMethod);
         }
 
+        /// <inheritdoc />
         public async Task<IDocument<T>> ExecuteTaskAsync<T>(IRestRequest request, CancellationToken token)
         {
             return await _browserTyped.ExecuteTaskAsync<T>(request, token);
         }
 
+        /// <inheritdoc />
         public async Task<IDocument<T>> ExecuteTaskAsync<T>(IRestRequest request)
         {
             return await _browserTyped.ExecuteTaskAsync<T>(request);
         }
 
+        /// <inheritdoc />
         public async Task<IDocument<T>> ExecuteGetTaskAsync<T>(IRestRequest request)
         {
             return await _browserTyped.ExecuteGetTaskAsync<T>(request);
         }
 
+        /// <inheritdoc />
         public async Task<IDocument<T>> ExecuteGetTaskAsync<T>(IRestRequest request, CancellationToken token)
         {
             return await _browserTyped.ExecuteGetTaskAsync<T>(request, token);
         }
 
+        /// <inheritdoc />
         public async Task<IDocument<T>> ExecutePostTaskAsync<T>(IRestRequest request)
         {
             return await _browserTyped.ExecutePostTaskAsync<T>(request);
         }
 
+        /// <inheritdoc />
         public async Task<IDocument<T>> ExecutePostTaskAsync<T>(IRestRequest request, CancellationToken token)
         {
             return await _browserTyped.ExecutePostTaskAsync<T>(request, token);
         }
 
+        /// <inheritdoc />
         public IDocument<T> Navigate<T>(string uri)
         {
             return _browserTyped.Navigate<T>(uri);
         }
 
+        /// <inheritdoc />
         public IDocument<T> Navigate<T>(Uri uri)
         {
             return _browserTyped.Navigate<T>(uri);
         }
 
+        /// <inheritdoc />
         public IDocument<T> Navigate<T>(string uri, Dictionary<string, string> headers)
         {
             return _browserTyped.Navigate<T>(uri, headers);
         }
 
+        /// <inheritdoc />
         public IDocument<T> Navigate<T>(Uri uri, Dictionary<string, string> headers)
         {
             return _browserTyped.Navigate<T>(uri, headers);
         }
 
+        /// <inheritdoc />
         public IDocument<T> Navigate<T>(string uri, Dictionary<string, string> headers, Dictionary<string, string> formData)
         {
             return _browserTyped.Navigate<T>(uri, headers, formData);
         }
 
+        /// <inheritdoc />
         public IDocument<T> Navigate<T>(Uri uri, Dictionary<string, string> headers, Dictionary<string, string> formData)
         {
             return _browserTyped.Navigate<T>(uri, headers, formData);
         }
 
+        /// <inheritdoc />
         public async Task<IDocument<T>> NavigateAsync<T>(string uri)
         {
             return await _browserTyped.NavigateAsync<T>(uri);
         }
 
+        /// <inheritdoc />
         public async Task<IDocument<T>> NavigateAsync<T>(Uri uri)
         {
             return await _browserTyped.NavigateAsync<T>(uri);
         }
 
+        /// <inheritdoc />
         public async Task<IDocument<T>> NavigateAsync<T>(string uri, Dictionary<string, string> headers)
         {
             return await _browserTyped.NavigateAsync<T>(uri, headers);
         }
 
+        /// <inheritdoc />
         public async Task<IDocument<T>> NavigateAsync<T>(Uri uri, Dictionary<string, string> headers)
         {
             return await _browserTyped.NavigateAsync<T>(uri, headers);
         }
-
+        
+        /// <inheritdoc />
         public async Task<IDocument<T>> NavigateAsync<T>(string uri, Dictionary<string, string> headers, Dictionary<string, string> formData)
         {
             return await _browserTyped.NavigateAsync<T>(uri, headers, formData);
         }
 
+        /// <inheritdoc />
         public async Task<IDocument<T>> NavigateAsync<T>(Uri uri, Dictionary<string, string> headers, Dictionary<string, string> formData)
         {
             return await _browserTyped.NavigateAsync<T>(uri, headers, formData);
         }
 
+        /// <inheritdoc />
         public IDocument<T> Submit<T>(string uri)
         {
             return _browserTyped.Submit<T>(uri);
         }
 
+        /// <inheritdoc />
         public IDocument<T> Submit<T>(Uri uri)
         {
             return _browserTyped.Submit<T>(uri);
         }
 
+        /// <inheritdoc />
         public IDocument<T> Submit<T>(string uri, Dictionary<string, string> formData)
         {
             return _browserTyped.Submit<T>(uri, formData);
         }
 
+        /// <inheritdoc />
         public IDocument<T> Submit<T>(Uri uri, Dictionary<string, string> formData)
         {
             return _browserTyped.Submit<T>(uri, formData);
         }
 
+        /// <inheritdoc />
         public IDocument<T> Submit<T>(string uri, Dictionary<string, string> formData, Dictionary<string, string> headers)
         {
             return _browserTyped.Submit<T>(uri, formData, headers);
         }
 
+        /// <inheritdoc />
         public IDocument<T> Submit<T>(Uri uri, Dictionary<string, string> formData, Dictionary<string, string> headers)
         {
             return _browserTyped.Submit<T>(uri, formData, headers);
         }
 
+        /// <inheritdoc />
         public IDocument<T> SubmitForm<T>(Form form)
         {
             return _browserTyped.SubmitForm<T>(form);
         }
 
+        /// <inheritdoc />
         public IDocument<T> SubmitForm<T>(Form form, Dictionary<string, string> headers)
         {
             return _browserTyped.SubmitForm<T>(form, headers);
         }
 
+        /// <inheritdoc />
         public async Task<IDocument<T>> SubmitAsync<T>(string uri)
         {
             return await _browserTyped.SubmitAsync<T>(uri);
         }
 
+        /// <inheritdoc />
         public async Task<IDocument<T>> SubmitAsync<T>(Uri uri)
         {
             return await _browserTyped.SubmitAsync<T>(uri);
         }
 
+        /// <inheritdoc />
         public async Task<IDocument<T>> SubmitAsync<T>(string uri, Dictionary<string, string> formData)
         {
             return await _browserTyped.SubmitAsync<T>(uri, formData);
         }
 
+        /// <inheritdoc />
         public async Task<IDocument<T>> SubmitAsync<T>(Uri uri, Dictionary<string, string> formData)
         {
             return await _browserTyped.SubmitAsync<T>(uri, formData);
         }
 
+        /// <inheritdoc />
         public async Task<IDocument<T>> SubmitAsync<T>(string uri, Dictionary<string, string> formData, Dictionary<string, string> headers)
         {
             return await _browserTyped.SubmitAsync<T>(uri, formData, headers);
         }
 
+        /// <inheritdoc />
         public async Task<IDocument<T>> SubmitAsync<T>(Uri uri, Dictionary<string, string> formData, Dictionary<string, string> headers)
         {
             return await _browserTyped.SubmitAsync<T>(uri, formData, headers);
         }
 
+        /// <inheritdoc />
         public async Task<IDocument<T>> SubmitFormAsync<T>(Form form)
         {
             return await _browserTyped.SubmitFormAsync<T>(form);
         }
 
+        /// <inheritdoc />
         public async Task<IDocument<T>> SubmitFormAsync<T>(Form form, Dictionary<string, string> headers)
         {
             return await _browserTyped.SubmitFormAsync<T>(form, headers);
