+++ BrowseSharp/Browsers/Core/BrowserCore.cs
@@ -23,7 +23,7 @@ using RestSharp.Deserializers;
 namespace BrowseSharp.Browsers.Core
 {
     /// <summary>
-    /// Headless browser implimentation that creates documents for each web request.
+    /// Headless browser core containing browser core attributes
     /// </summary>
     public class BrowserCore : IBrowserCore
     {
